feat(eval): T 격자를 실수로 받고 반올림을 절단기와 맞춘다
**정수 격자가 성기다.** CoVoST2 test 는 문장이 평균 9.1 어절이라 T 2->3 에서 조각 수가
4.79 -> 3.03 으로 건너뛴다. 우리 점수 격자의 S60(k 3.91)·S80(2.93)이 바로 그 구간이고,
실측 LAAL 1,379~1,826ms 에는 비교군 점이 하나도 없어서 S90·S95 의 우위(+3.3, +3.8)가
450ms 짜리 보간 위에 떠 있었다. `--t-grid` 가 실수를 받게 하고 2.5·3.5·4.5·5 를 넣어
k 간격을 0.2~0.4 로 고른다. 조건 이름은 `syntax_T2.5` 처럼 붙는다.

**반올림도 어긋나 있었다.** `coarsen` 은 파이썬 기본 `round`(은행가 반올림)를 쓰고
절단기는 `round_half_up` 을 쓴다. 5어절을 T=2 로 나누면 2.5 가 나오는데 이쪽은 2,
저쪽은 3 이었다 — 짧은 문장이 많은 코퍼스에서는 T=2 조건이 통째로 달라진다. 주석이
"절단기와 같은 규칙" 이라고 적은 대로 `round_half_up` 으로 맞춘다.

테스트 둘을 더한다: 실수 T 가 정수 사이에 떨어지는지, 그리고 예산 일치 검사에 .5 가
나오는 값(2.5·3.5)을 넣어 두 반올림이 갈리면 바로 걸리게 한다.

+++ core/meaning_segmentator/autoseg/scoring/bleu_eval.py
@@ -315,7 +315,10 @@ def main() -> int:
                         "과 정렬 파일 이름에 쓴다. X→en 트랙은 de/ja/zh 를 준다")
     p.add_argument("--src-spaced", type=int, default=1,
                    help="measured_profile.json 이 없을 때 쓸 소스 띄어쓰기 여부")
-    p.add_argument("--t-grid", type=int, nargs="+", default=[4, 6, 8, 12])
+    # **실수를 받는다.** 정수만 받으면 T 2 와 3 사이를 못 메우는데, 짧은 코퍼스에서는
+    # 거기서 조각 수가 4.8 에서 3.0 으로 건너뛴다 (CoVoST2 test 평균 9.1 어절 실측).
+    # 조건 이름은 `syntax_T2.5` 처럼 붙는다.
+    p.add_argument("--t-grid", type=float, nargs="+", default=[4, 6, 8, 12])
     p.add_argument("--no-auto-t", action="store_true",
                    help="우리 프롬프트의 `auto_T*` 조건을 만들지 않는다 — 지연 노브를 점수 임계값"
                         "(`--score-grid`)으로만 낼 때. **`--t-grid` 는 그대로 줘야 한다**: 비교군의 "
